interactive playground; interactive; tidied up examples
playground.html gives an immediate rendering of the script you type.

draw_interactive_svg renders interactive drawings: the variables time,
mousex and mousey are added to the scope that the drawing is evaluated
in.

The examples on index.html are displayed more nicely, with code next to
canvas and svg renderings.