add ARIA labels to just about everything
Elements drawn in diagrams get automatically generated ARIA labels.
For some things, these might inadvertently include spoilers, such as the
coordinates of a point of intersection or the radius of a circle.
I'm not sure what to do to make these accessible - just saying "circle"
won't be much use, but you'd need some understanding of the semantics of
the diagram to say something like "circle inscribed in square".

Maybe a command `description(text)` to override the label?

A "no spoilers" mode, which avoids revealing any information not given
on screen, would compromise accessibility but prevent accidentally
spoiling questions.