add aria_mode and description()
if aria mode is VERBOSE, then information not explicitly displayed, such
as coordinates and angles, is included in ARIA labels. Otherwise, only
names of elements (such as "circle" and "dot") and labelled points are used.

If all three points of an angle are labelled, e.g. A B and C, its label
will be "angle A B C". Otherwise it'll just be "angle".

You can use `verbose` or `nospoilers` to set the aria mode. The default
is `nospoilers`.

The description(text) function sets the aria label for the affected
element to the given text.